requirements.txt
Flask==3.0.0
Flask-Cors==4.0.0
Flask-Limiter==3.5.0
requests==2.31.0
python-dotenv==1.0.0
gunicorn==21.2.0